Add files via upload
this script searches the fda data dashboard by firm name that the user enters. it outputs links to the firm profile where the user can view warning letters that firm has received. the script also contains a section that outputs links directly to the warning letters, however it does not consistently work due to inconsistencies in how firm names are listed on the fda data dashboard.